Fix serialization for `undefined`
When serializing an object that contains `undefined` as a value,
the current packer code incorrectly serializes the value as `null`
(0xc0). `null` and `undefined` are not the same thing, and should
not be serialized into the same binary representation.

Strangely, the unpacker actually does support unpacking `undefined`
values using the binary representation 0xc1.

There was an old pull request (#4) that seems to address this along
with a bunch of other seemingly unrelated changes. The GitHub UI
seems to think that the PR was merged manually, but the code that
fixes this issue never actually made it into the master branch as
far as I can tell.

Another option is to omit any keys with `undefined` values. This is
what `JSON.stringify` does as `JSON` does not support `undefined`
natively. When deserialized, the semantics are mostly correct, as
accessing a key that doesn't exist will give you `undefined`. This
is slightly more efficient, I suppose and maintain the exact same
behavior between binarypack and JSON serialization which is a good
thing. However, you can observe the difference with `Object.keys`,
`Object.hasOwnProperty`, etc.